Move The Buffer shenanigans into the ShapeComponentIds in GltfUtils
